The massive Code Documentation PR (#175)
* PathItem Documentation

* Renamed PathItem

* blockedVertex doc

* fixed typos

* Chunk docs

* GriddedMap docs

* Region docs

* GridDesign docs

* Refactor of relative address int>double (for offset grids)

* cleaned typos

* Map docs

* fixed lSense @see link

* paritial MapLayer docs - finish after merge

* refactored distacne from long>double and AngleDistance docs

* Waypoint docs

* fixed misc typos

* bossable docs

* connectable docs

* contactable docs

* directable docs

* Drawable docs

* driveable docs

* leader.disscussMerge sig change

* some clean up

* started adjusting @see to FQN

* convert @see to FQN

* groupable docs

* instructable docs

* joinable docs

* LSensible docs

* notifiable docs

* organisable docs

* promotable docs

* transferable docs

* updateable docs

* RemoteMember docs

* RemoteLeader docs

* misc cleanups

* fixed typos and javadoc errors

* MapLayer Reversion for merge

* maplayer test

* maplayer revert

+++ mcomp_Java/src/common/datatypes/AngleDistance.java
@@ -1,29 +1,50 @@
-/**
- * 
- */
 package common.datatypes;
 
 /**
+ * Immutable Tuple for holding an Angle (theta) and distance representation of a Waypoint
+ * 
  * @author David Avery 15823926
+ * @author Stephen Pope 15836791
+ * 
+ * @version 1.0
+ * @since 2018-04-07
+ * 
+ * @see common.datatypes.Waypoint
  *
  */
 public class AngleDistance {
   double theta;
-  long distance;
+  double distance;
 
   /**
+   * Constructor
+   * 
+   * @param t Angle (in degrees) offset from 0 (x axis) in a counter-clockwise aspect
+   * @param d Distance (in centimetres) from the origin to the date point
    * 
    */
-  AngleDistance(double t, long d) {
+  AngleDistance(double t, double d) {
     theta = t;
     distance = d;
   }
 
+  /**
+   * Access method for getting theta value
+   * 
+   * @return Theta value in degrees as a double
+   * 
+   */
   double getTheta() {
     return theta;
   }
 
-  long getDistance() {
+  /**
+   * Access method for getting distance value
+   * 
+   * @return Distance value in centimetres as a double
+   * 
+   */
+  double getDistance() {
     return distance;
   }
